# xergon-marketplace/docs/openapi.yaml
openapi: "3.0.2"
info:
  title: Xergon Relay API
  version: "1.0.0"
  description: |
    Decentralized AI inference relay for the Xergon Network. OpenAI-compatible
    API that routes inference requests to GPU providers on the Ergo blockchain.
    No signup required -- authenticate with an Ergo keypair.
  contact:
    name: Xergon Network
    url: https://degens.world
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://relay.xergon.gg
    description: Production relay
...
security:
  - hmacSignature: []
paths:
  /v1/chat/completions:
    post:
      tags: [Inference]
      summary: Create chat completion
      description: |
        OpenAI-compatible chat completion endpoint. Supports streaming (SSE)
        and non-streaming responses. Routes to the best available provider
        based on PoNW score, latency, and model availability.
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Successful completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream (stream=true)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: X-Xergon-Signature
      description: |
        HMAC-SHA256 signature of (request_body + timestamp).
        Requires three headers:
        - X-Xergon-Timestamp: Unix timestamp (seconds)
        - X-Xergon-Signature: HMAC-SHA256(body + timestamp, private_key)
        - X-Xergon-Public-Key: User's Ergo public key (hex)
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
              description: Machine-readable error type
              enum:
                - invalid_request
                - unauthorized
                - forbidden
                - not_found
                - rate_limit_error
                - internal_error
                - service_unavailable
            message:
              type: string
              description: Human-readable error message
            code:
              type: integer
              description: HTTP status code
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Model identifier (e.g., "qwen3.5-32b")
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        max_tokens:
          type: integer
          default: 1024
        temperature:
          type: number
          default: 0.7
          minimum: 0
          maximum: 2
        top_p:
          type: number
          default: 1
        stream:
          type: boolean
          default: false
          description: Enable Server-Sent Events streaming
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                $ref: "#/components/schemas/ChatMessage"
              finish_reason:
                type: string
                enum: [stop, length, content_filter]
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              type: invalid_request
              message: "Missing required field: model"
              code: 400
    Unauthorized:
      description: Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              type: unauthorized
              message: "Invalid HMAC signature"
              code: 401
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              type: rate_limit_error
              message: "Rate limit exceeded (trial tier: 5 req/min)"
              code: 429
    ServiceUnavailable:
      description: No providers available
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              type: service_unavailable
              message: "No providers available for the requested model"
              code: 503
...